% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figumem.R
\name{figumem_item}
\alias{figumem_item}
\title{Create a figumem-item}
\usage{
figumem_item(
  objn = 20,
  mode = "classic",
  challenge = 1,
  seed = sample(1:999999999, 1),
  ntrys = 100
)
}
\arguments{
\item{objn}{The number of frame-emblem units to be used. Can be any natural number between
1 and 20. Default is 20.}

\item{mode}{The answering format of the figumem item. Can be either "classic",
"classic_paper", "draw" or "draw_random". Default is "classic".}

\item{challenge}{The difficulty of the item. Must be either 1, 2 or 3. 1 is the lowest,
3 is the highest difficulty. Default is 1.}

\item{seed}{The seed number for item production. Default is using a random seed number.}

\item{ntrys}{The maximum number of tries until an item is either successfully produced or
the production is cancelled. Default is 100.}
}
\value{
A figumem item as a list containing images.
}
\description{
This function creates a figumem item. for more detailed information and examples please
refer to the figumem-manual available under
https://raw.githubusercontent.com/Jendryczko/figumem/master/figumem_Manual.pdf
}
