% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figumem.R
\name{figumem_retrieve_online}
\alias{figumem_retrieve_online}
\title{Retrieve the figumem stimuli from Github}
\usage{
figumem_retrieve_online()
}
\value{
figumem_stimuli: A list of the figumem stimuli.
}
\description{
This function retrieves the png-data-files containing the figumem stimuli
from Github and loads them into a list called "figumem_stimuli". This list
is required for the production of figumem items, i. e. for the function
"figumem_item" to work. The first 20 elements of the list contain the 20 frames
that are used for items of the lowest difficulty. Elemts 21 to 40 contain the frames
for items of medium and the hardest difficulty. Elements 41 to 60 contain the emblems
in a small format for the lowest and medium difficulty. Elements 61 to 80 contain the emblems
in a small format for the hardest difficulty. Elements 81 to 100 contain the emblems
in a large format for the lowest and medium difficulty. Elements 101 to 120 contain the emblems
in a large format for the hardest difficulty.
The function will only work if the computer is connected to the internet!
}
