% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figumem.R
\name{figumem_write}
\alias{figumem_write}
\title{Save a figumem item}
\usage{
figumem_write(item, name = "figural_memory_item", format = "png")
}
\arguments{
\item{item}{The figumem item to be saved. This argument has no default.}

\item{name}{The name for the figumem item. Default is "figural_memory_item". If you
want to save the item somewhere different than your current working directory you
can also specify the complete path here.}

\item{format}{The format of the figumem item. Must be either "png", "jpg" or "pdf". Default
is "png". "pdf" is only recommended for items with modes "classic_paper", "draw" and "draw_random".}
}
\description{
This function saves a figumem item in either png-, jpg- or pdf-format.
For more detailed information and examples please refer to the figumem-manual available under
https://raw.githubusercontent.com/Jendryczko/figumem/master/figumem_Manual.pdf
}
